Extract evidence pack inspection projection helper
